# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is user communication service.",
        "title": "Communication Graph",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/graph": {
            "get": {
                "description": "return graph nodes links",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Graph",
                "operationId": "graph",
                "responses": {
                    "200": {
                        "description": "status",
                        "schema": {
                            "$ref": "#/definitions/mapper.JournalRecord"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/record": {
            "post": {
                "description": "add new communication node",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "record"
                ],
                "summary": "Record",
                "operationId": "record",
                "parameters": [
                    {
                        "description": "account info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mapper.JournalRecord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "mapper.JournalRecord": {
            "type": "object",
            "properties": {
                "source_id": {
                    "type": "string"
                },
                "target_id": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        }
    }
}